Unify data generation
Unify the data generation task in the Rakefile.  Also remove the need
for the file name by removing the multiple-inclusion guard.  No file
should be including any of these files more than once (and, in fact,
only one file should include any of them for the whole project) anyway,
and it’s better to be notified if this would be the case.

We also extract the common constants that define the table sizes.